conda refactor - kamangir/bolt#420

+++ .abcli/conda.sh
@@ -4,7 +4,7 @@ function roofAI_conda() {
     local task=$(abcli_unpack_keyword $1 help)
 
     if [ "$task" == "help" ]; then
-        abcli_show_usage "roofAI conda create_env [validate]" \
+        abcli_show_usage "roofAI conda create_env [validate,~recreate]" \
             "create conda environment."
         abcli_show_usage "roofAI conda validate" \
             "validate conda environment."
@@ -13,9 +13,17 @@ function roofAI_conda() {
 
     if [ "$task" == "create_env" ]; then
         local options=$2
+        local do_recreate=$(abcli_option_int "$options" recreate 1)
         local do_validate=$(abcli_option_int "$options" validate 0)
 
-        abcli_conda create_env name=roofAI
+        local environment_name=roofAI
+
+        if [[ "$do_recreate" == 0 ]] && [[ $(abcli_conda exists $environment_name) == 1 ]]; then
+            abcli_eval - conda activate $environment_name
+            return
+        fi
+
+        abcli_conda create_env name=$environment_name
 
         pip3 install pymysql==0.10.1
         pip3 install -U albumentations[imgaug]
